ci: refactor release & branch management strategy (#67)
**Related Issue(s):** #

We're currently experiencing issues triggering releases in an automated
fashion. I believe that this is due to the fact that release-please is
configured to only run on merges to `main`:


https://github.com/source-cooperative/data.source.coop/blob/2aafefe93bf2d958d0c3bd868a5ab491b0b995b1/.github/workflows/release.yaml#L1-L5

This is a non-default branch, as we follow a gitflow model for branch
management. Reading the release-please docs[^1], the tool appears to be
built to follow a trunk model for branch management as there is no
acknowledgement of a separate production branch or any indication that
the tool would work with such a flow.

**Proposed Changes:**

1. Point CICD workflows to the `main` branch, eliminating the notion of
`dev` branch
2. Add conventional commit check to PR titles via
[`ytanikin/pr-conventional-commits@1.4.0`](https://github.com/ytanikin/pr-conventional-commits)
3. Update PR template to allow contributors to better describe their
changes

**Actions to be performed after merge:**

1. Delete `main` branch
2. Rename `dev` to `main`
4. Configure repo to utilize squash merges as described here:
https://github.com/ytanikin/pr-conventional-commits?tab=readme-ov-file#configuring-squash-merging

**PR Checklist:**

- [ ] This PR is made against the dev branch (all proposed changes
except releases should be against dev, not main).
- [ ] This PR has **no** breaking changes.
- [ ] This PR contains only one commit.
- [ ] I have updated or added new tests to cover the changes in this PR.
- [ ] All tests are passing.
- [ ] This PR affects the [Source Cooperative Frontend &
API](https://github.com/source-cooperative/source.coop),
      and I have opened issue/PR #XXX to track the change.

[^1]:
https://github.com/googleapis/release-please-action?tab=readme-ov-file#how-release-please-works